Refactor: UseCase 어노테이션 추가
