+++ proma/src/main/java/ai/platform/proma/annotation/UseCase.java
@@ -1,0 +1,16 @@
+package ai.platform.proma.annotation;
+
+import org.springframework.core.annotation.AliasFor;
+import org.springframework.stereotype.Component;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Component
+public @interface UseCase {
+
+    @AliasFor(annotation = Component.class)
+    String value() default "";
+}
